Create the 'PamtramEditorSourceSectionMatcherPage'
First version of the page that allows to visualize source section
matches in a source model. Currently, only a source section viewer is
implemented.
Add source model selection combo

Add a combo-box that allows to select the project of which the source
sections shall be matched. The entries of the combo-box are set to the
projects in the 'Source' folder of the project (if it has the pamtram
nature assigned).
Display selected source model in matcher page

If the user selects a model in the combo box, it is loaded and displayed
in the appropriate tree viewer so that the matches can be visualized.