Implemented method 'mapSections()'
This method runs the transformation to the point where the source
sections have been mapped and returns them. This is only a preliminary
implementation that can be used by the 'SourceSectionMatcherPage' in the
pamtram editor. In the future, this feature should probably be realized
by storing each performed transformation in the pamtram and using this
information to visualize the matching source sections.