Support ecore-based metamodels in the editor
The editor now explicitly supports ecore-based metamodels that are not
backed by generated Java code. Therefore, before populating the editor,
ePackages involved in the pamtram model are determined and it is checked
if they are already present in the epackage registry. If this is not the
case, the 'metamodel' folder of the active project is searched for ecore
models that could hold the ePackage. If a matching ePackage is found, it
is registered. As a result, ecore metamodels should not have to be
registered via the epsilon feature 'Right click -> register ePackage'
any longer.