Improved the pre-transformation check if all ePackages are registered
Models are now loaded via 'platform' instead of 'file' URIs. Thus, they
can be loaded, even if not all packages are found. After the packages
have been checked, the model is only reloaded if packages had to be
registered manually. Thus, the models normally are only loaded once
(instead of twice with the old approach).