The content adapter now incorporates multiple child content adapters
This should allow for a better feature separation between differnt
adapters while at the same time keeping the concept of a single adapter
for better performance.
Additionally, this adds a new adapter that listens to changes in
collections of AttValueModifierSets and asks the user if he wants to
update other collections that reference the same AttValueModifierSets
with the same changes automatically.