Fix a bug for connection hints with different target classes
Model connection hints may specify different target classes. Up to now,
the same model connection path was used for every instance to connect
via the same model connection hint. Now, the eClass of the target
containers are also considered and if they are not equal, a new model
connection hint is calculated.