Fix for SourceSectionMatcher with modeled empty references
If a SourceSectionReference is modeled without any target values, the
intended meaning is that a part of the source model can only be machted
if it does not contain any target elements for this particular
reference.

Until now, the interpretation by the transformation was that a part of
the source model could be matched independent of it containing any
target elements for the reference or not.

This fix changes the interpretation to the desired behavior.