Create infrastructure to find more specific library entry
Check the AttributeMappings if there is one that provides a new library
path. If there is one, use the new path to get a more specific library
entry. Up to now, the new entry is not checked for parameter equivalence
with the old one and the old entry is not exchanged by a new one so that
the result of a transformation is the same as with the previous commit.