Use abstract mapping for bool/double in/out
Create an abstract source section 'viObject Base' that covers the label,
x, y, .... This is used by abstract hint groups to create the labels and
text definition. Thus, each of the for mappings for bool/double in/out
now simply imports the abstract hint groups which significantly reduces
the number of mapping hints necessary.