Support overwriting of mapping hints
If a hint group extends another (abstract) hint group and defines a
mapping hint that affects an element of the target section that is also
affected by a hint of the abstract hint group, the hint of the abstract
hint group is not copied to the concrete hint group. Consequently, the
concrete hint 'overwrites' the abstract hint.