Create interface 'IAmbiguityResolvingStrategy'
This interface needs to be implemented by strategies that provide a
possibility to resolve ambiguities that arise during the execution of a
generic transformation. In the future, the only existing strategy ('let
the user decide') will implement this. By making use of the strategy
pattern, we can simply pass a list of concrete strategies to the
'GenericTransformationRunner' that can then simply consult every given
strategy without knowing about implementation details.

This should greatly simplify the addition of new strategies (e.g. the
consulting of previous choices that are stored in a 'transformation
model').