Add 'gentrans.transformation' plugin dependency
Due to the introduction of the strategy pattern for ambiguity resolving
in the 'transformation' plugin, the transformation plugin does not know
any longer about the 'selectors' plugin as this will in the future just
implement a special ambiguity resolving strategy.

For this to work, the 'selectors' plugin needs however to depend on the
'transformation' plugin to be able to implement the
'IAmbiguityResolvingStrategy' interface.