Adapt the TargetModelRegistry to work with the genlibrary
Due to a call to 'eObject.eResource()' during the insertion of a library
entry by the genlibrary, the strategy to only add all contents to the
target resources in the end does not work and throws an NPE. Thus, we
now directly add the contents to the target resources (and create them
if necessary). At the end of the transformation, all resources are
simply saved.