Use the new 'SetViewerMouseListener' in the editor pages
This should ensure that no 'SelectionEvents' get lost: Before, if an
element in a tree viewer was selected via a right click, the focus would
not be set to the new object under certain conditions. Consequently, the
'createNewChild' actions would still be suitable for the old selected
element and not for the new one.