Create a method 'isLocalCondition' for type 'ComplexConditon'
This allows to check if a condition is 'local', i.e. it is based on the
SourceSection that is referenced by the mapping containing the
condition, or 'global', i.e. it is based on any other SourceSection.
This is necessary as 'global' and 'local' conditions need to be handled
differently.