Preliminary support for 'Reference Match Specs'
A 'Reference Match Spec' is an optional list of References that can be
defined for a LocalDynamicSourceElement. If present, it describes a path
how to get to the relevant attribute instances based on the root element
of the Section. This can be used to prevent ambiguous matches especially
in case of self-referencing section definitions.