Create a new feature project for standalone applications
This feature project is not used as classical feature project. However,
it declares all plugins that are necessary for an application of the
GenTrans plugins outside eclipse. That way, we can make use of the
'export feature' option to export all those plugins at once which can
then be used as classpath by standalone projects.