Fix for the setup file
Up to now, we just installed the
'org.eclipse.epp.package.modeling.feature' feature group which was
equivalent to installing an 'Eclipse 4 Modeling'. Strangely, this does
not seem to be the case anymore (e.g. the EMF SDK is no longer
installed). Thus, we manually install everything that is necessary to
get a version equivalent to an 'Eclipse 4 Modeling'.