Linking: Do not link cross references with neither value nor selector
Up to now, if no 'value' was specified for a TargetSectionCrossReference
(and no ReferenceTargetSelector), all instances of all compatible
classes were considered potential target elements. This however often
lead to many user queries with a multitude of potential options. In
order to reduce these queries, cross references with neither value nor
selector specified are not considered any longer in the linking process.