GenTrans: Start refactoring the matching process
Instead of checking CrossReferences directly (equivalent to
CompositeReferences), defer checking and create associated
'MatchingDependencies' instead that are checked later.

This should simplify handling of cyclic references between Sections.
Additionally, this makes the matching process easier to understand
overall as we do not need to jump to other Sections during the matching
of one Section.